requirements.txt
django<2.1
psycopg2<2.8
djangorestframework<3.8
